# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Cucumber REST API",
    "version": "v1"
  },
  "paths": {
...
    "/raven/1.0/import/execution/cucumber/multipart": {
      "post": {
        "operationId": "importExecutionCucumberMultipart",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "info": {
                    "$ref": "#/components/schemas/FormDataBodyPart"
                  },
                  "result": {
                    "$ref": "#/components/schemas/FormDataBodyPart"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportExecutionResponse"
                }
              }
            },
            "description": "Cucumber test result successfully imported"
          }
        },
        "summary": "Import cucumber test result (multipart)",
        "tags": [
          "Import"
        ]
      }
    },
    "/raven/1.0/import/execution/junit": {
      "post": {
        "operationId": "importExecutionJunit",
        "parameters": [
          {
            "in": "query",
            "name": "projectKey",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "testExecKey",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "testPlanKey",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "testEnvironments",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "revision",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "fixVersion",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "$ref": "#/components/schemas/FormDataBodyPart"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportExecutionResponse"
                }
              }
            },
            "description": "JUnit test result successfully imported"
          }
        },
        "summary": "Import JUnit test result",
        "tags": [
          "Import"
        ]
      }
    },
    "/raven/1.0/import/execution/junit/multipart": {
      "post": {
        "operationId": "importExecutionJunitMultipart",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "info": {
                    "$ref": "#/components/schemas/FormDataBodyPart"
                  },
                  "file": {
                    "$ref": "#/components/schemas/FormDataBodyPart"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportExecutionResponse"
                }
              }
            },
            "description": "JUnit test result successfully imported"
          }
        },
        "summary": "Import JUnit test result (multipart)",
        "tags": [
          "Import"
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "BodyPart": {
        "properties": {
          "contentDisposition": {
            "$ref": "#/components/schemas/ContentDisposition"
          },
          "entity": {
            "type": "object"
          },
          "headers": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "properties": {
              "empty": {
                "type": "boolean"
              }
            },
            "type": "object"
          },
          "mediaType": {
            "$ref": "#/components/schemas/MediaType"
          },
          "messageBodyWorkers": {
            "$ref": "#/components/schemas/MessageBodyWorkers"
          },
          "parameterizedHeaders": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/ParameterizedHeader"
              },
              "type": "array"
            },
            "properties": {
              "empty": {
                "type": "boolean"
              }
            },
            "type": "object"
          },
          "parent": {
            "$ref": "#/components/schemas/MultiPart"
          },
          "providers": {
            "$ref": "#/components/schemas/Providers"
          }
        },
        "type": "object"
      },
      "ContentDisposition": {
        "properties": {
          "creationDate": {
            "format": "date-time",
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "modificationDate": {
            "format": "date-time",
            "type": "string"
          },
          "parameters": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "readDate": {
            "format": "date-time",
            "type": "string"
          },
          "size": {
            "format": "int64",
            "type": "integer"
          },
          "type": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "FormDataBodyPart": {
        "properties": {
          "content": {
            "type": "object"
          },
          "contentDisposition": {
            "$ref": "#/components/schemas/ContentDisposition"
          },
          "entity": {
            "type": "object"
          },
          "fileName": {
            "type": "string"
          },
          "formDataContentDisposition": {
            "$ref": "#/components/schemas/FormDataContentDisposition"
          },
          "headers": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "properties": {
              "empty": {
                "type": "boolean"
              }
            },
            "type": "object"
          },
          "mediaType": {
            "$ref": "#/components/schemas/MediaType"
          },
          "messageBodyWorkers": {
            "$ref": "#/components/schemas/MessageBodyWorkers"
          },
          "name": {
            "type": "string"
          },
          "parameterizedHeaders": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/ParameterizedHeader"
              },
              "type": "array"
            },
            "properties": {
              "empty": {
                "type": "boolean"
              }
            },
            "type": "object"
          },
          "parent": {
            "$ref": "#/components/schemas/MultiPart"
          },
          "providers": {
            "$ref": "#/components/schemas/Providers"
          },
          "simple": {
            "type": "boolean"
          },
          "value": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "FormDataContentDisposition": {
        "properties": {
          "creationDate": {
            "format": "date-time",
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "modificationDate": {
            "format": "date-time",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "parameters": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "readDate": {
            "format": "date-time",
            "type": "string"
          },
          "size": {
            "format": "int64",
            "type": "integer"
          },
          "type": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ImportExecutionResponse": {
        "description": "Represents the response after importing a test execution, containing the test execution issue details",
        "properties": {
          "testExecIssue": {
            "$ref": "#/components/schemas/TestExecIssue"
          }
        },
        "type": "object"
      },
      "MediaType": {
        "properties": {
          "parameters": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "subtype": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "wildcardSubtype": {
            "type": "boolean"
          },
          "wildcardType": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "MessageBodyWorkers": {
        "type": "object"
      },
      "MultiPart": {
        "properties": {
          "bodyParts": {
            "items": {
              "$ref": "#/components/schemas/BodyPart"
            },
            "type": "array"
          },
          "contentDisposition": {
            "$ref": "#/components/schemas/ContentDisposition"
          },
          "entity": {
            "type": "object"
          },
          "headers": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "properties": {
              "empty": {
                "type": "boolean"
              }
            },
            "type": "object"
          },
          "mediaType": {
            "$ref": "#/components/schemas/MediaType"
          },
          "messageBodyWorkers": {
            "$ref": "#/components/schemas/MessageBodyWorkers"
          },
          "parameterizedHeaders": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/ParameterizedHeader"
              },
              "type": "array"
            },
            "properties": {
              "empty": {
                "type": "boolean"
              }
            },
            "type": "object"
          },
          "parent": {
            "$ref": "#/components/schemas/MultiPart"
          },
          "providers": {
            "$ref": "#/components/schemas/Providers"
          }
        },
        "type": "object"
      },
...
      "ParameterizedHeader": {
        "properties": {
          "parameters": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "value": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "Providers": {
        "type": "object"
      },
...
      "TestExecIssue": {
        "description": "Represents a test execution issue",
        "properties": {
          "id": {
            "description": "The unique identifier of the test execution issue",
            "type": "string"
          },
          "key": {
            "description": "The key of the test execution issue",
            "type": "string"
          },
          "self": {
            "description": "The self URL of the test execution issue",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  }
}